Created valence_boss_bar crates and created the BossBar entity
