+++ crates/valence_boss_bar/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "valence_boss_bar"
+description = "Boss bar API for Valence"
+version = "0.1.0"
+edition = "2021"
+readme = "README.md"
+keywords = ["minecraft", "bossbar", "api"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+valence_core.workspace = true
+valence_network.workspace = true
+valence_entity.workspace = true
+uuid.workspace = true
+bitfield-struct.workspace = true
+bevy_app.workspace = true
+bevy_ecs.workspace = true
