+++ crates/valence_boss_bar/README.md
@@ -1,0 +1,3 @@
+# valence_boss_bar
+
+Valence's Boss Bar API
