+++ crates/valence_boss_bar/src/packet.rs
@@ -1,0 +1,27 @@
+use uuid::Uuid;
+use valence_core::{protocol::{Encode, Decode, Packet, packet_id}, text::Text};
+
+use crate::{BossBarColor, BossBarDivision, BossBarFlags};
+
+#[derive(Clone, Debug, Encode, Decode, Packet)]
+#[packet(id = packet_id::BOSS_BAR_S2C)]
+pub struct BossBarS2c {
+    pub id: Uuid,
+    pub action: BossBarAction,
+}
+
+#[derive(Clone, PartialEq, Debug, Encode, Decode)]
+pub enum BossBarAction {
+    Add {
+        title: Text,
+        health: f32,
+        color: BossBarColor,
+        division: BossBarDivision,
+        flags: BossBarFlags,
+    },
+    Remove,
+    UpdateHealth(f32),
+    UpdateTitle(Text),
+    UpdateStyle(BossBarColor, BossBarDivision),
+    UpdateFlags(BossBarFlags),
+}
